指标数据权限

+++ svr/svr-ehr-basic/src/main/java/com.yihu.ehr/geography/service/GeographyDictService.java
@@ -35,7 +35,7 @@ public class GeographyDictService extends BaseJpaService<GeographyDict,XGeograph
 	}
 
     public GeographyDict findById(String id) {
-        return geographyDictRepository.findOne(Long.valueOf(id));
+        return geographyDictRepository.findOne(Integer.valueOf(id));
     }
 
     public GeographyDict findByName(String name) {
